ONE-8994: Add migration test, update schemas to match data model

+++ apps/core-server/src/endpoint/trust_list_publication/dto.rs
@@ -109,7 +109,7 @@ pub(crate) struct GetTrustListPublicationResponseRestDTO {
     pub role: TrustListPublicationRoleRestEnum,
     #[from(with_fn_ref = "map_raw_str_opt")]
     pub content: Option<String>,
-    pub sequence_number: i64,
+    pub sequence_number: u32,
     #[from(with_fn_ref = "map_raw_str")]
     pub metadata: String,
     pub organisation_id: OrganisationId,
